demo/requirements.txt
accelerate==0.28.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
Authlib==1.3.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.1
build==1.1.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
cryptography==42.0.5
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
datasets==2.18.0
dill==0.3.8
distlib==0.3.8
distro==1.9.0
dulwich==0.21.7
emoji==2.10.1
fastjsonschema==2.19.1
filelock==3.13.1
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
grpcio==1.62.1
grpcio-health-checking==1.62.1
grpcio-tools==1.62.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
httpx-sse==0.4.0
idna==3.6
importlib_metadata==7.0.2
installer==0.7.0
isodate==0.6.1
jaraco.classes==3.3.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==24.3.1
langchain==0.1.12
langchain-cli==0.0.21
langchain-community==0.0.28
langchain-core==0.1.32
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.29
libmagic==1.0
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
openai==1.14.2
orjson==3.9.15
packaging==23.2
pexpect==4.9.0
pillow==10.2.0
pkginfo==1.10.0
platformdirs==4.2.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.0
Pygments==2.17.2
pyparsing==3.1.2
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.2.7
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.1
rdflib==6.0.2
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.0
scipy==1.12.0
sentence-transformers==2.6.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
sse-starlette==1.8.2
starlette==0.36.3
streamlit==1.32.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tokenizers==0.15.2
toml==0.10.2
tomlkit==0.12.4
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
trove-classifiers==2024.3.3
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.18
uvicorn==0.23.2
validators==0.22.0
virtualenv==20.25.1
weaviate-client==4.5.4
wrapt==1.16.0
xattr==1.1.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
